requirements.txt
pyfiglet
colorama
discord_webhook
pycryptodome